requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
python-dotenv==1.0.0
Werkzeug==2.2.3